+++ data/src/main/java/com/semicolon/data/remote/response/notice/NoticeWriter.kt
@@ -3,7 +3,7 @@ package com.semicolon.data.remote.response.notice
 import com.google.gson.annotations.SerializedName
 
 data class NoticeWriter (
-    @SerializedName("id") val writerId: Long,
+    @SerializedName("id") val writerId: Int,
     @SerializedName("name") val writerName: String,
     @SerializedName("profile_url") val profileUrl: String
 )
